v11.0.0rc3
- Update dependencies
- Update builder config
- Update README
