+++ build.js
@@ -14,17 +14,7 @@ execSync('rm -rf dist')
 execSync('mkdir dist')
 
 const buildTime = (new Date()).toISOString()
-const longComment = [
-  '/**',
-  ` * ${pkg.name}@${pkg.version}`,
-  ` * built with esbuild at: ${buildTime}`,
-  ` * repository: ${pkg.repository.url}`,
-  ` * maintainer: ${pkg.author}`,
-  ` * License: ${pkg.license}`,
-  '**/'
-].join('\n')
-
-const shortComment = [
+const comment = [
   `// ${pkg.name}@${pkg.version}, by ${pkg.author}`,
   `built with esbuild at ${buildTime}`,
   `published under ${pkg.license} license`
@@ -46,7 +36,7 @@ const esmVersion = {
   mainFields: ['module'],
   outfile: 'dist/bella.esm.js',
   banner: {
-    js: longComment
+    js: comment
   }
 }
 buildSync(esmVersion)
@@ -58,7 +48,7 @@ const cjsVersion = {
   mainFields: ['main'],
   outfile: 'dist/cjs/bella.js',
   banner: {
-    js: longComment
+    js: comment
   }
 }
 buildSync(cjsVersion)
@@ -83,7 +73,7 @@ const iifeVersion = {
   globalName: 'bella',
   outfile: 'dist/bella.min.js',
   banner: {
-    js: shortComment
+    js: comment
   }
 }
 buildSync(iifeVersion)
